Improved search 'gemeente', added simplified gemeentenamen, simplified html of gemeente search result list
Refactored the code to store gemeente alternative names only in the
backend, and let the frontend work in a generic way.
That way the knowledge only needs to be in one place.
Also added additional alternative names. To make sure only the original
name shows up in the list, the results are put in a Set (as both names
will possible be in the list, e.g. original and alternative name).
Names are also sorted to show the in alphabetical order.

Closes #94 Added additional alternative 'gemeentenamen'.
Closes #129 All search results in a single ul element.